Support distributed training

+++ training/transformer/train.py
@@ -24,6 +24,7 @@ from training.datasets.convert_lieder import convert_lieder, lieder_train_index
 from training.datasets.convert_primus import convert_primus_dataset, primus_train_index
 from training.run_id import get_run_id
 from training.transformer.data_loader import label_names, load_dataset
+from training.transformer.distribute import Distribute
 from training.transformer.metrics import HomrTrainer
 from training.transformer.mix_datasets import mix_training_sets
 
@@ -114,6 +115,8 @@ def _check_datasets_are_present(selected_datasets: list[str]) -> list[str]:
 def train_transformer(
     fp32: bool = False, resume: str = "", smoke_test: bool = False, fine_tune: bool = False
 ) -> None:
+    distribute = Distribute()
+
     number_of_epochs = 35
     if smoke_test:
         number_of_epochs = 10
@@ -125,23 +128,25 @@ def train_transformer(
     if resume:
         resume_from_checkpoint = os.path.join(git_root, checkpoint_folder, resume)
     elif os.path.exists(os.path.join(git_root, checkpoint_folder)):
-        shutil.rmtree(os.path.join(git_root, checkpoint_folder))
+        if distribute.is_rank0():
+            shutil.rmtree(os.path.join(git_root, checkpoint_folder))
+
+    dataset_index = [lieder_train_index, grandstaff_train_index, primus_train_index]
+    if distribute.is_rank0():
+        _check_datasets_are_present(dataset_index)
+    distribute.barrier()
 
     if smoke_test:
         number_of_files = -1
         train_index = load_and_mix_training_sets(
-            _check_datasets_are_present(
-                [lieder_train_index, grandstaff_train_index, primus_train_index]
-            ),
+            dataset_index,
             [1.0, 1.0, 1.0],
             number_of_files,
         )
     else:
         number_of_files = -1
         train_index = load_and_mix_training_sets(
-            _check_datasets_are_present(
-                [lieder_train_index, grandstaff_train_index, primus_train_index]
-            ),
+            dataset_index,
             [1.0, 1.0, 1.0],
             number_of_files,
         )
@@ -167,7 +172,7 @@ def train_transformer(
         save_strategy="epoch",
         learning_rate=1e-5 if fine_tune else 1e-4,
         optim="adamw_torch_fused",
-        gradient_accumulation_steps=4,
+        gradient_accumulation_steps=max(1, 4 // distribute.get_world_size()),
         per_device_train_batch_size=batch_size,
         per_device_eval_batch_size=batch_size // 2,
         num_train_epochs=number_of_epochs,
@@ -202,6 +207,7 @@ def train_transformer(
 
     if os.path.exists(model_destination):
         eprint("Model already exists", model_destination)
+        distribute.destroy()
         return
 
     try:
@@ -215,13 +221,17 @@ def train_transformer(
             train_dataset=datasets["train"],
             eval_dataset=datasets["validation"],
             callbacks=callbacks,
+            distribute=distribute,
         )
 
         trainer.train(resume_from_checkpoint=resume_from_checkpoint)
     except KeyboardInterrupt:
         eprint("Interrupted")
-    torch.save(model.state_dict(), model_destination)
-    eprint(f"Saved model to {model_destination}")
+    if distribute.is_rank0():
+        torch.save(model.state_dict(), model_destination)
+        eprint(f"Saved model to {model_destination}")
+    distribute.barrier()
+    distribute.destroy()
 
 
 if __name__ == "__main__":
